refactor: reorganize components and update import paths
- Moved MaterialSymbol and TransitionLink to a new `common` subdirectory
- Updated import paths across multiple files to reflect new component location
- Restructured register page components under `components/pages/register`
- Maintained existing functionality while improving project structure
