+++ app/components/pages/register/step2.tsx
